Extract fs extensions to separate module
