+++ packages/brookjs-cli/src/index.ts
@@ -1,8 +1,9 @@
-import * as glob from './glob';
 import * as eslint from './eslint';
+import * as glob from './glob';
+import * as fs from './fs';
 
 export * from './services';
 export * from './cli';
 export { default as create } from './create';
 
-export { glob, eslint };
+export { glob, eslint, fs };
