+++ packages/brookjs-cli/src/services/index.ts
@@ -1,6 +1,5 @@
 import * as project from './project';
 import * as prettier from './prettier';
 
-export { default as fs } from './fs';
 export { project, prettier };
 export { default as WebpackService } from './webpack';
